+++ modules/home-manager/_hyprland/looknfeel.nix
@@ -13,7 +13,7 @@ in
   wayland.windowManager.hyprland.settings = {
     general = {
       gaps_in = 5;
-      gaps_out = 10;
+      gaps_out = 5;
 
       border_size = 2;